+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file is used to list changes made in each version of the users cookbook.
 
-## Unreleased
+## 8.1.0 - *2021-08-26*
 
 - Add `system` property to `users_manage` resource
 - Add `system` property to user json test data